Move function protos to proto/fn
I'm not considering this a breaking change because in practice we
replicate these protos to each language SDK, so I don't expect any (or
many) folks to be importing them directly.

My motivation for moving them is that they're no longer used only for
apiextensions (composition). We'll use the same proto for ops.
