+++ Earthfile
@@ -136,7 +136,7 @@ go-modules-tidy:
   FROM +go-modules
   CACHE --id go-build --sharing shared /root/.cache/go-build
   COPY generate.go .
-  COPY --dir apis/ cmd/ internal/ test/ .
+  COPY --dir apis/ proto/ cmd/ internal/ test/ .
   RUN go mod tidy
   RUN go mod verify
   SAVE ARTIFACT go.mod AS LOCAL go.mod
@@ -149,7 +149,7 @@ go-generate:
   COPY +kubectl-setup/kubectl /usr/local/bin/kubectl
   COPY generate.go buf.yaml buf.gen.yaml .
   COPY --dir cluster/crd-patches cluster/crd-patches
-  COPY --dir hack/ apis/ internal/ .
+  COPY --dir hack/ apis/ proto/ internal/ .
   RUN go generate -tags 'generate' .
   # TODO(negz): Can this move into generate.go? Ideally it would live there with
   # the code that actually generates the CRDs, but it depends on kubectl.
@@ -159,6 +159,7 @@ go-generate:
     --output=yaml > /tmp/patched.yaml \
     && mv /tmp/patched.yaml cluster/crds/pkg.crossplane.io_deploymentruntimeconfigs.yaml
   SAVE ARTIFACT apis/ AS LOCAL apis
+  SAVE ARTIFACT proto/ AS LOCAL proto
   SAVE ARTIFACT internal/ AS LOCAL internal
   SAVE ARTIFACT cluster/crds AS LOCAL cluster/crds
   SAVE ARTIFACT cluster/meta AS LOCAL cluster/meta
@@ -180,7 +181,7 @@ go-build:
     SET ext = ".exe"
   END
   CACHE --id go-build --sharing shared /root/.cache/go-build
-  COPY --dir apis/ cmd/ internal/ .
+  COPY --dir apis/ proto/ cmd/ internal/ .
   RUN go build -o crossplane${ext} ./cmd/crossplane
   RUN sha256sum crossplane${ext} | head -c 64 > crossplane${ext}.sha256
   RUN go build -o crank${ext} ./cmd/crank
@@ -212,7 +213,7 @@ go-build-e2e:
   ARG CGO_ENABLED=0
   FROM +go-modules
   CACHE --id go-build --sharing shared /root/.cache/go-build
-  COPY --dir apis/ internal/ test/ .
+  COPY --dir apis/ proto/ internal/ test/ .
   RUN go test -c -o e2e ./test/e2e
   SAVE ARTIFACT e2e
 
@@ -220,7 +221,7 @@ go-build-e2e:
 go-test:
   FROM +go-modules
   CACHE --id go-build --sharing shared /root/.cache/go-build
-  COPY --dir apis/ cmd/ internal/ .
+  COPY --dir apis/ proto/ cmd/ internal/ .
   RUN go test -covermode=count -coverprofile=coverage.txt ./...
   SAVE ARTIFACT coverage.txt AS LOCAL _output/tests/coverage.txt
 
@@ -233,9 +234,10 @@ go-lint:
   CACHE --id go-build --sharing shared /root/.cache/go-build
   RUN curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin ${GOLANGCI_LINT_VERSION}
   COPY .golangci.yml .
-  COPY --dir apis/ cmd/ internal/ test/ .
+  COPY --dir apis/ proto/ cmd/ internal/ test/ .
   RUN golangci-lint run --fix
   SAVE ARTIFACT apis AS LOCAL apis
+  SAVE ARTIFACT proto AS LOCAL proto
   SAVE ARTIFACT cmd AS LOCAL cmd
   SAVE ARTIFACT internal AS LOCAL internal
   SAVE ARTIFACT test AS LOCAL test
@@ -420,7 +422,7 @@ ci-codeql:
   END
   COPY --dir +ci-codeql-setup/codeql /codeql
   CACHE --id go-build --sharing shared /root/.cache/go-build
-  COPY --dir apis/ cmd/ internal/ .
+  COPY --dir apis/ proto/ cmd/ internal/ .
   RUN /codeql/codeql database create /codeqldb --language=go
   RUN /codeql/codeql database analyze /codeqldb --threads=0 --format=sarif-latest --output=go.sarif --sarif-add-baseline-file-info
   SAVE ARTIFACT go.sarif AS LOCAL _output/codeql/go.sarif
