+++ internal/xfn/cached/cached_runner_test.go
@@ -32,8 +32,8 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 
-	fnv1 "github.com/crossplane/crossplane/apis/apiextensions/fn/proto/v1"
 	"github.com/crossplane/crossplane/internal/xfn/cached/proto/v1alpha1"
+	fnv1 "github.com/crossplane/crossplane/proto/fn/v1"
 )
 
 type FunctionRunnerFn func(ctx context.Context, name string, req *fnv1.RunFunctionRequest) (*fnv1.RunFunctionResponse, error)
