+++ internal/xfn/function_runner_test.go
@@ -31,9 +31,9 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/test"
 
-	fnv1 "github.com/crossplane/crossplane/apis/apiextensions/fn/proto/v1"
-	fnv1beta1 "github.com/crossplane/crossplane/apis/apiextensions/fn/proto/v1beta1"
 	pkgv1 "github.com/crossplane/crossplane/apis/pkg/v1"
+	fnv1 "github.com/crossplane/crossplane/proto/fn/v1"
+	fnv1beta1 "github.com/crossplane/crossplane/proto/fn/v1beta1"
 )
 
 var _ fnv1.FunctionRunnerServiceClient = &BetaFallBackFunctionRunnerServiceClient{}
