+++ proto/fn/v1/run_function.proto
@@ -22,7 +22,7 @@ package apiextensions.fn.proto.v1;
 import "google/protobuf/duration.proto";
 import "google/protobuf/struct.proto";
 
-option go_package = "github.com/crossplane/crossplane/apis/apiextensions/fn/proto/v1";
+option go_package = "github.com/crossplane/crossplane/proto/fn/v1";
 
 // A FunctionRunnerService is a Composition Function.
 service FunctionRunnerService {
