+++ proto/fn/v1beta1/zz_generated_run_function.proto
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-// Generated from apis/apiextensions/fn/proto/v1/run_function.proto by ./hack/duplicate_proto_type.sh. DO NOT EDIT.
+// Generated from proto/fn/v1/run_function.proto by ./hack/duplicate_proto_type.sh. DO NOT EDIT.
 
 syntax = "proto3";
 
@@ -24,7 +24,7 @@ package apiextensions.fn.proto.v1beta1;
 import "google/protobuf/duration.proto";
 import "google/protobuf/struct.proto";
 
-option go_package = "github.com/crossplane/crossplane/apis/apiextensions/fn/proto/v1beta1";
+option go_package = "github.com/crossplane/crossplane/proto/fn/v1beta1";
 
 // A FunctionRunnerService is a Composition Function.
 service FunctionRunnerService {
